Fix and add tests for SSF queueing per sender
*wipes forehead*

well well well, this was a doozie.

This patch fixes the issue where running SSF across multiple machines would
not result in any messages coming in. Turns Out™ this was a pretty obvious
bug with the implementation—SSF requires one pending queue and expected
timestamp **per** sender. In other words, if P' and P'' were in a network,
and P' broadcasted M', it would see its own M' and increment its single expected
timestamp to 1, causing the next message from P'' (with timestamp 0) to be queued
indefinitely.

This fix replaces the single expected timestamp with one per sender.