Replace deprecated Obs.create with unsafeCreate
As of RxJava 1.2.7 [1] the Observable.create method taking a function
of a subscriber has been deprecated and the overloads taking either
a SyncOnSubscribe instance, an AsyncOnSubscribe instance, or a
function of an rx.Emitter [2] are now preferred. This change here
updates UdpBroadcast to use the Observable.unsafeCreate while waiting
for #20 to replace it with one of the new overloads.

  [1]:https://github.com/ReactiveX/RxJava/releases/tag/v1.2.7
  [2]:http://reactivex.io/RxJava/1.x/javadoc/rx/Emitter.html