Remove transient ByteBuffer from Sender
Kryo wasn't reconstructing the object correctly due to the transient field. It
was remaining empty because there wasn't anywhere to re-wrap the buffer during
Kryo's deserialization process. This change removes the field altogether,
replacing it with wrapping the buffer everywhere it is used (a silly amount of
allocations, I know).

I plan on reducing the allocations once the dust settles.