Queue up all messages that should be delivered and deliver them in one go
This works fine where this is all on a single thread—the previous implementation
had the delivery using the wrong timestamps (as CausalOrder#prepare would be
called before the clocks had been adjusted). This implementation adjusts the
clocks correctly and delivers all of the messages in one go with the right timestamps.