Remove reference checks from equals implementations
The check can essentially be replaced with false and everything will
still work fine. Sure, it's _slower_, but who really compares an
object to itself enough times for it to matter? I imagine the common
case is comparison to objects that are not itself. (I know, I know,
what a wild idea, right?)